exclude . and .. from pathname expansions
Plan 9 doesn't include . and .. in directory listings (as expected as
they are more like path traversal utilities and have no business being
listed as directory entries).

Unix however originally implemented them as directory entries and most
Unix-like systems include them in what readdir() returns.

Ironically, a bug in an early version of ls caused all files whose name
started with . to not be listed when the intention was only to skip .
and ..

As a result, it became a "feature". So now Unix systems are littered
with files whose name starts with "." and that are meant to be hidden
(by ls, or not expanded by globs unless explicitly asked for). To be
useful to Unix users, a Unix port of rc has to follow that same
convention.

However including . and .. is never desirable. rm had to be patched to
work around the disastrous effects of rm -rf .* in shells that include
them. A chmod -R 644 .* still has disastrous effects.

A few shells (Forsyth, pdksh and derivatives, fish, zsh) always exclude
. and .. from glob expansions. es, when built with SHOW_DOT_FILES on (to
behave like Plan 9 rc), skips  . and ..

The POSIX specification of sh is currently being modified to allow that
behaviour (and maybe even mandate it in the future).
http://austingroupbugs.net/view.php?id=1228

This commit implements it for this shell.